Add files via upload
2022 Final Year Group Project for a Domestic Helper Application.

Purpose of this project is to allow users to see services that regisitered employees offer domestically. These services are based on location (in this project, Counties in Kent), price and review rating. Users who are interested in booking a service must register on the website to book.

Bookings are done by Date/Time, and checks are in place to prevent double bookings.

Written in the following stack:

HTML5
CSS
JS/ jQuery (sweetalert2 for alerts)
PHP8
MySQL8

Improvements in the next revision in this project will include:

- Optimise for Mobile Development
- Improved Calendar functionality
- Implement Facebook/Google Login
- More Dynamic Features